fix(sensor): 修复传感器配置和通信协议问题
- 修改PIR触发边沿从下降沿到上升沿，实现人体接近检测
- 添加V0值范围限制防止定时器溢出
- 增强updateV0函数增加数据验证和范围检查
- 在sensorInit中初始化本地地址为ZXBEE_ADDR_SENSOR_C
- 扩展ZXbee通信协议支持16位地址长度
- 更新ZXbee帧格式从7字节到9字节头部结构
- 修改ZXBee地址定义从8位扩展到16位数值
- 添加对night和arm命令的处理支持
- 修复字符串比较中的字段名兼容性问题
- 优化ZXBee命令解析增加更精确的条件判断

+++ zstack-2.4.0-1.4.0x/Projects/zstack/Samples/common/zxbee-inf.h
@@ -1,8 +1,10 @@
 #ifndef __ZXBEE_INF_H__
 #define __ZXBEE_INF_H__
 
+#include "hal_types.h"
+
 void ZXBeeInfInit(void);
-void ZXBeeInfSetLocalAddr(uint8 addr);
+void ZXBeeInfSetLocalAddr(uint16 addr);
 void ZXBeeInfSend(char *p, int len);
 void ZXBeeInfRecv(char *pkg, int len);
 #endif
